I wasn't happy with how prescriptive the ARM template and Build Script were. Because they always prefixed the website name with "safe-" and suffixed it with "-web". So I made some changes so that I can call my site whatever I want. I've also added a proper appName environment variable to set the app Service Name.
I also pulled some of the variables out of the ARM template and made them into parameters so that the names can be cleaned in the script before passing them to the ARM template. This will make cleaning the names easier in the future for anybody who wants to customize names. It is now checking for proper name for storage which didn't always work before if the name "environment" was too long. I've also added filtering for the resourceGroup name with has to be alfanumeric or _-.()

It is still possible to omit the appName and Environment variables and get the GUID name as before.